Fixed problems with the within(min, val, max) function. Was using ints and duplicating functionality of bound(min, val, max).

+++ robot/inc/util_math.h
@@ -7,7 +7,7 @@ typedef struct {
 } pt;
 
 float distance(int x1, int y1, int x2, int y2);
-int within(int min, int val, int max);
+float within(float min, float val, float max);
 float bound(float min, float val, float max);
 
 #endif
